% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_database_management.R
\name{alter_retention_policy}
\alias{alter_retention_policy}
\title{alter_retention_policy}
\usage{
alter_retention_policy(con, rp_name, db, duration, replication,
  default = FALSE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{rp_name}{The name of the retention policy.}

\item{db}{The name of the database.}

\item{duration}{Determines how long InfluxDB keeps the data.}

\item{replication}{The number of data nodes.}

\item{default}{logical. If TRUE, the new retention policy is the default retention policy
for the database.}
}
\value{
A list of server responses.
}
\description{
Alter retention policy
}
\details{
This function is a convenient wrapper for modifying a retention policy on a
given database by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
