% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_helper.R
\name{check_srv_comm}
\alias{check_srv_comm}
\title{method to check the server communication results
function is not exported}
\usage{
check_srv_comm(x)
}
\arguments{
\item{x}{httr::POST response}
}
\description{
method to check the server communication results
function is not exported
}
\keyword{internal}
