% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_authentication_and_authorization.R
\name{create_user}
\alias{create_user}
\title{create_user}
\usage{
create_user(con, username, password)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{username}{Sets the username.}

\item{password}{Sets the password.}
}
\description{
Create user
}
\details{
This function is a convenient wrapper for creating a new user
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
