% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_diagnostics.R
\name{show_stats}
\alias{show_stats}
\alias{show_diagnostics}
\title{Show diagnostics and stats.}
\usage{
show_stats(con)

show_diagnostics(con)
}
\arguments{
\item{con}{An \code{influx_connection} object
(s. \code{\link{influx_connection}}).}
}
\value{
A data.frame frame.
}
\description{
These functions calls \code{influx_query} to receive some stats.
}
\section{Warning}{
 \code{show_stats} might take some time.
}

\seealso{
\code{\link[influxdbr]{influx_connection}}
}
