% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_database_management.R
\name{drop_measurement}
\alias{drop_measurement}
\title{drop_measurement}
\usage{
drop_measurement(con, db, measurement)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Sets the measurement to be dropped.}
}
\value{
A list of server responses.
}
\description{
Drop measurement
}
\details{
This function is a convenient wrapper for dropping an entire measurement
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
