% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb_drop_data.R
\name{drop_series}
\alias{drop_series}
\title{drop_series}
\usage{
drop_series(con, db, id = NULL, from = NULL, where = NULL)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{id}{Sets the series ID.}

\item{from}{Sets a specific measurement.}

\item{where}{Apply filter on tag key values.}
}
\value{
A list of server responses.
}
\description{
Drop series
}
\details{
This function is a convenient wrapper for dropping an individual series
within a measurement that match given tags by calling \code{influx_query}
with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/query_language/data_exploration.html}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

