% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_authentication_and_authorization.R
\name{drop_user}
\alias{drop_user}
\title{drop_user}
\usage{
drop_user(con, username)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{username}{Specifies the user to be dropped.}
}
\description{
Drop user
}
\details{
This function is a convenient wrapper for dropping a user from an influxdb
server by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/administration/authentication_and_authorization/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

