% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_helper.R
\name{get_precision_divisor}
\alias{get_precision_divisor}
\title{method to transform precision divisor
function is not exported}
\usage{
get_precision_divisor(x)
}
\arguments{
\item{x}{character}
}
\description{
method to transform precision divisor
function is not exported
}
\keyword{internal}
