% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_helper.R
\name{httr_GET}
\alias{httr_GET}
\title{method send the httr::GET command
function is not exported}
\usage{
httr_GET(con, query = NULL, endpoint)
}
\arguments{
\item{con}{influx_connection object}

\item{endpoint}{api endpoint}
}
\value{
httr::GET response
}
\description{
method send the httr::GET command
function is not exported
}
\keyword{internal}
