% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_helper.R
\name{httr_POST}
\alias{httr_POST}
\title{method send the httr::POST command
function is not exported}
\usage{
httr_POST(con, query = NULL, body = NULL, endpoint)
}
\arguments{
\item{con}{influx_connection object}

\item{endpoint}{api endpoint}
}
\value{
httr::POST response
}
\description{
method send the httr::POST command
function is not exported
}
\keyword{internal}
