% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_connection}
\alias{influx_connection}
\title{influx_connection}
\usage{
influx_connection(host = NULL, port = NULL, user = "user",
  pass = "pass", group = NULL, verbose = FALSE,
  config_file = "~/.influxdb.cnf")
}
\arguments{
\item{host}{hostname}

\item{port}{numerical. port number}

\item{user}{username}

\item{pass}{password}

\item{group}{groupname}

\item{verbose}{logical. Provide additional details?}

\item{config_file}{The configuration file to be used if \code{groupname} is
specified.}
}
\description{
Create an influxdb_connection object
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/}
A group file looks like:
[groupname]
host=localhost
port=8086
user=username
pass=password
}

