% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_main.R
\name{influx_ping}
\alias{influx_ping}
\title{Ping an influxdb server}
\usage{
influx_ping(con)
}
\arguments{
\item{con}{An \code{influx_connection} object (s. \code{\link{influx_connection}}).}
}
\value{
A tibble with server information.
}
\description{
This function pings an influxdb server (e.g. for connection testing)
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr]{influx_connection}}
}
