% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_post.R
\name{influx_post}
\alias{influx_post}
\title{send POST to an InfluxDB server}
\usage{
influx_post(con, db = NULL, query = "")
}
\arguments{
\item{con}{An \code{influx_connection} object (s. \code{\link{influx_connection}}).}

\item{db}{Sets the target database for the query.}

\item{query}{The InfluxDB query to be sent.}
}
\value{
A tibble or NULL
}
\description{
This function sends POST to an InfluxDB server. It is not
exported and only used for some helper functions within this package.
}
\references{
\url{https://influxdb.com/}
}
\keyword{internal}
