% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_query_xts}
\alias{influx_query_xts}
\title{influx_query_xts}
\usage{
influx_query_xts(con, db = NULL, query = "SELECT * FROM measurement",
  timestamp_format = "default", verbose = FALSE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the name of the database.}

\item{query}{The influxdb query to be sent.}

\item{timestamp_format}{Sets the timestamp format
("default" (=UTC), "n", "u", "ms", "s", "m", "h").}

\item{verbose}{logical. Provide additional details?}
}
\value{
A list of influx_series objects.
        Each influx_series object will have name, query, tags and values as members.
        name is usually the measurement name from which the data is fetched.
        values is a xts object representing the time series data
        tags is a list of tags associated with the series
}
\description{
Query an influxdb server
}
\author{
mvadu (\email{info@adystech.com})
}
\references{
\url{https://influxdb.com/docs/v0.9/concepts/api.html}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr2]{influx_connection}}
}

