% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_main.R
\name{influx_select}
\alias{influx_select}
\title{influx_select}
\usage{
influx_select(con, db, field_keys, rp = NULL, measurement, where = NULL,
  group_by = NULL, limit = NULL, slimit = FALSE, offset = NULL,
  order_desc = FALSE, return_xts = TRUE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the name of the database.}

\item{field_keys}{Specifies the fields to be selected.}

\item{rp}{The name of the retention policy.}

\item{measurement}{Sets the name of the measurement.}

\item{where}{Apply filter on tag key values.}

\item{group_by}{The group_by clause in InfluxDB is used not only for
grouping by given values, but also for grouping by given time buckets.}

\item{limit}{Limits the number of the n oldest points to be returned.}

\item{slimit}{logical. Sets limiting procedure (slimit vs. limit).}

\item{offset}{Offsets the returned points by the value provided.}

\item{order_desc}{logical. Change sort order to descending.}

\item{return_xts}{logical. Sets the return type. If set to TRUE, xts objects
are returned, FALSE gives data.frames.}
}
\value{
A list of xts or data.frame objects.
}
\description{
This function is a convenient wrapper for selecting data from a measurement
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/query_language/data_exploration/}
}

