% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_write_df.R
\name{influx_write_df}
\alias{influx_write_df}
\title{Write a data.frame object to an InfluxDB server}
\usage{
influx_write_df(con, db, measurement = NULL, x, time_col = NULL,
  tag_cols = NULL, rp = NULL, precision = c("s", "ns", "u", "ms", "m",
  "h"), consistency = c(NULL, "one", "quroum", "all", "any"),
  max_points = 5000, use_integers = FALSE)
}
\arguments{
\item{con}{An \code{influx_connection} object (s. \code{\link{influx_connection}}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Sets the name of the measurement.}

\item{x}{The data.frame object to write to an InfluxDB server.}

\item{time_col}{A character scalar naming the time index column.}

\item{tag_cols}{A character vector naming tag columns.}

\item{rp}{Sets the target retention policy for the write. If not present the
default retention policy is used.}

\item{precision}{Sets the precision of the supplied Unix time values
("s", "ns", "u", "ms", "m", "h"). If not present timestamps are assumed to be
in seconds.}

\item{consistency}{Set the number of nodes that must confirm the write.
If the requirement is not met the return value will be partial write
if some points in the batch fail, or write failure if all points in the batch
fail.}

\item{max_points}{Defines the maximum points per batch (defaults to 5000).}

\item{use_integers}{Should integers (instead of doubles) be written if present?}
}
\value{
A list of server responses.
}
\description{
This function writes a data.frame object to an InfluxDB server.
Columns may contain times and both tag and field values.
Columnnames of the data.frame object are used as InfluxDB's tag and field keys.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr]{influx_connection}}
}
