% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_authentication_and_authorization.R
\name{revoke_privileges}
\alias{revoke_privileges}
\title{revoke_privileges}
\usage{
revoke_privileges(con, privilege = c("READ", "WRITE", "ALL"), db, username)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{privilege}{Specifies the user privilege.}

\item{db}{Sets the target database.}

\item{username}{Specifies the user.}
}
\description{
Revoke database privileges
}
\details{
This function is a convenient wrapper for revoking database privileges to an
existing user by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
