% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_databases}
\alias{show_databases}
\title{show_databases}
\usage{
show_databases(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\value{
A character vector containing the database names.
}
\description{
Show databases
}
\details{
This function is a convenient wrapper for showing all databases
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
