% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb_diagnostics.R
\name{show_diagnostics}
\alias{show_diagnostics}
\title{show_diagnostics}
\usage{
show_diagnostics(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\value{
A list with diagnostics.
}
\description{
Show diagnostics
}
\details{
This function is a convenient wrapper for showing diagnostics
by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

