% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_series}
\alias{show_series}
\title{Show series}
\usage{
show_series(con, db, measurement = NULL, where = NULL, verbose = TRUE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Query a specific measurement.}

\item{where}{Apply filter on tag key values.}

\item{verbose}{logical. Occasionally, an influxdb server might be queried without
specifying 'measurement' and/or 'where' which could result in a large number of
individual series. Due to the function's purpose to result one single data.frame
containing all series, the merging process of mulitple data.frames with unequal
colnames potentially take some time. If \code{verbose} is set to \code{TRUE},
the user is informed about an evtl. long running process and can interrupt the
function call. The server result will be returned unpolished in this case.
Defaults to TRUE.}
}
\value{
A list of data.frame objects.
}
\description{
This function is a convenient wrapper for showing all series with distinct
key-value pairs by calling \code{influx_query} with
the corresponding query. Series can be shown for a specific meausurement with
the (\code{measurement}) parameter and can also be filtered by tag key values with
the \code{where} clause.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
