% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_diagnostics.R
\name{show_stats}
\alias{show_stats}
\title{Show stats}
\usage{
show_stats(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\value{
A list of data.frame objects.
}
\description{
This function is a convenient wrapper for showing stats
by calling \code{influx_query} with the corresponding query.
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
