% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_tag_values}
\alias{show_tag_values}
\title{Show tag values}
\usage{
show_tag_values(con, db, measurement = NULL, key)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Query a specific measurement.}

\item{key}{The key to be queried.}
}
\value{
A character vector containing tag values.
}
\description{
This function is a convenient wrapper for showing the unique set of
tag values for each measurement, for a given tag key by calling
\code{influx_query} with the corresponding query.
Tag values can be filtered by a specific measurement (\code{measurement}).
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
