% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_authentication_and_authorization.R
\name{show_users}
\alias{show_users}
\title{show_users}
\usage{
show_users(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\description{
Show users
}
\details{
This function is a convenient wrapper for listing all existent users
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
