% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_database_management.R
\name{create_database}
\alias{create_database}
\title{create_database}
\usage{
create_database(con, db)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database to create.}
}
\value{
A list of server responses.
}
\description{
Create database
}
\details{
This function is a convenient wrapper for creating a database
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/query_language/database_management/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
