% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb.R
\name{create_database}
\alias{create_database}
\title{create_database}
\usage{
create_database(con, db)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database to create.}
}
\value{
A list of server responses.
}
\description{
Create databases
}
\details{
This function is a convenient wrapper for creating a database
by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/introduction/getting_started.html}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

