% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_database_management.R
\name{drop_database}
\alias{drop_database}
\title{Drop databases}
\usage{
drop_database(con, db)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database to drop.}
}
\value{
A list of server responses.
}
\description{
This function is a convenient wrapper for dropping a database
by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
