% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb_drop_data.R
\name{drop_measurement}
\alias{drop_measurement}
\title{drop_measurement}
\usage{
drop_measurement(con, db, measurement)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Sets the measurement to be dropped.}
}
\value{
A list of server responses.
}
\description{
Drop measurement
}
\details{
This function is a convenient wrapper for dropping an entire measurement
by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/query_language/data_exploration.html}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

