% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_database_management.R
\name{drop_retention_policy}
\alias{drop_retention_policy}
\title{drop_retention_policy}
\usage{
drop_retention_policy(con, rp_name, db)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{rp_name}{The name of the retention policy.}

\item{db}{The name of the database.}
}
\value{
A list of server responses.
}
\description{
Delete retention policy
}
\details{
This function is a convenient wrapper for deleting a retention policy on a
given database by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/query_language/database_management/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

