% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_database_management.R
\name{drop_series}
\alias{drop_series}
\title{drop_series}
\usage{
drop_series(con, db, id = NULL, measurement = NULL, where = NULL)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{id}{Sets the series ID.}

\item{measurement}{Sets a specific measurement.}

\item{where}{Apply filter on tag key values.}
}
\value{
A list of server responses.
}
\description{
Drop series
}
\details{
This function is a convenient wrapper for dropping an individual series
within a measurement that match given tags by calling \code{influx_query}
with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
