% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_authentication_and_authorization.R
\name{grant_privileges}
\alias{grant_privileges}
\title{grant_privileges}
\usage{
grant_privileges(con, privilege = c("READ", "WRITE", "ALL"), db, username)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{privilege}{Specifies the user privilege.}

\item{db}{Sets the target database.}

\item{username}{Specifies the user.}
}
\description{
Grant database privileges
}
\details{
This function is a convenient wrapper for granting database privileges to an
existing user by calling \code{influx_query} with the corresponding query.
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/administration/authentication_and_authorization/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
