% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_connection}
\alias{influx_connection}
\title{influx_connection}
\usage{
influx_connection(scheme = "http", host = "localhost", port = 8086,
  user = "user", pass = "pass", path = "/", group = NULL,
  verbose = FALSE, config_file = "~/.influxdb.cnf")
}
\arguments{
\item{scheme}{The scheme to use, either http or https. Defaults to http}

\item{host}{Hostname of the InfluxDB server. Defaults to localhost}

\item{port}{numerical. port number of the InfluxDB server. Defaults to the default port 8086}

\item{user}{username The username to use. Defaults to "user"}

\item{pass}{password The password to use. Defaults to "pass".}

\item{path}{The prefix path on which the influxdb is running. Can be useful in proxy situations.}

\item{group}{groupname The group to use within the group config file}

\item{verbose}{logical. Provide additional details?}

\item{config_file}{The configuration file to be used if \code{groupname} is
specified.}
}
\description{
Create an influxdb_connection object
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/}
A group file looks like:
[groupname]
host=localhost
port=8086
user=username
pass=password
}

