% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_ping}
\alias{influx_ping}
\title{influx_ping}
\usage{
influx_ping(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\value{
A list of server information.
}
\description{
Ping an influxdb server
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/concepts/api.html}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr]{influx_connection}}
}

