% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_query}
\alias{influx_query}
\title{influx_query}
\usage{
influx_query(con, db = NULL, query = "SELECT * FROM measurement",
  timestamp_format = "default", return_xts = TRUE, verbose = FALSE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the name of the database.}

\item{query}{The influxdb query to be sent.}

\item{timestamp_format}{Sets the timestamp format
("default" (=UTC), "n", "u", "ms", "s", "m", "h").}

\item{return_xts}{logical. Sets the return type. If set to TRUE, xts objects
are returned, FALSE gives data.frames.}

\item{verbose}{logical. Provide additional details?}
}
\value{
A list of xts or data.frame objects.
}
\description{
Query an influxdb server
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/concepts/api.html}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr]{influx_connection}}
}

