% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_query.R
\name{influx_query}
\alias{influx_query}
\title{Query an InfluxDB server}
\usage{
influx_query(con, db = NULL, query = "SELECT * FROM measurement",
  timestamp_format = c("n", "u", "ms", "s", "m", "h"),
  tags_as_factors = TRUE, chunked = FALSE, handler = NULL,
  verbose = FALSE)
}
\arguments{
\item{con}{An \code{influx_connection} object
(s. \code{\link{influx_connection}}).}

\item{db}{Sets the target database for the query.}

\item{query}{The InfluxDB query to be sent.}

\item{timestamp_format}{Sets the timestamp format ("n", "u", "ms", "s", "m",
"h").}

\item{tags_as_factors}{If TRUE, convert tag columns to factors automatically.}

\item{chunked}{Either FALSE or an integer. If FALSE, series are not requested
in streamed batches. If an integer is provided, responses will be chunked
by series or by every \code{chunked} points. If \code{handler} is not
supplied chunks are aggregated into a data.frame.}

\item{handler}{A function receiving a list of parsed json objects when
\code{chunked} it not FALSE. \code{influx_query} returns a list of values
returned by \code{handler}. If handler returns NULL the value is omitted
from the output. Supply \code{identity} to retrieve raw parsed json.}

\item{verbose}{If TRUE print log messages.}
}
\value{
A list a data.frame object. Empty query returns NULL.
}
\description{
This functions queries an InfluxDB server.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
