% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_select}
\alias{influx_select}
\title{influx_select}
\usage{
influx_select(con, db, value, rp = NULL, from, where = NULL,
  group_by = NULL, limit = NULL, offset = NULL, return_xts = TRUE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the name of the database.}

\item{value}{Sets the name of values to be selected.}

\item{rp}{The name of the retention policy.}

\item{from}{Sets the name of the measurement.}

\item{where}{Apply filter on tag key values.}

\item{group_by}{The group_by clause in InfluxDB is used not only for
grouping by given values, but also for grouping by given time buckets.}

\item{limit}{Limits the number of the n oldest points to be returned.}

\item{offset}{Offsets the returned points by the value provided.}

\item{return_xts}{logical. Sets the return type. If set to TRUE, xts objects
are returned, FALSE gives data.frames.}
}
\value{
A list of xts or data.frame objects.
}
\description{
influx_select
}

