% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb.R
\name{influx_write}
\alias{influx_write}
\title{influx_write}
\usage{
influx_write(con, db, xts, measurement = NULL, precision = "s",
  max_points = 5000, digits = 5, performance = F)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{The name of the database.}

\item{xts}{The xts object to write to an influxdb server.}

\item{measurement}{The name of the measurement.}

\item{precision}{Specifies the timestamp format ("default", "n", "u", "ms", "s", "m", "h").}

\item{max_points}{Defines the maximum points per batch.}

\item{digits}{How many digits are to be used.}

\item{performance}{logical. Print performance measurements.}
}
\value{
A list of server responses.
}
\description{
Write an xts object to an influxdb server
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/guides/writing_data.html}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr]{influx_connection}}
}

