% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_main.R
\name{influx_write}
\alias{influx_write}
\title{influx_write}
\usage{
influx_write(con, db, xts, measurement = NULL, rp = NULL,
  precision = c("s", "n", "u", "ms", "m", "h"), consistency = c(NULL, "one",
  "quroum", "all", "any"), max_points = 5000, use_integers = FALSE)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the write.}

\item{xts}{The xts object to write to an influxdb server.}

\item{measurement}{Sets the name of the measurement.}

\item{rp}{Sets the target retention policy for the write. If not present the
default retention policy is used.}

\item{precision}{Sets the precision of the supplied Unix time values
("n", "u", "ms", "s", "m", "h"). If not present timestamps are assumed to be
in nanoseconds. Currently only "s" is supported.}

\item{consistency}{Set the number of nodes that must confirm the write.
If the requirement is not met the return value will be partial write
if some points in the batch fail, or write failure if all points in the batch
fail.}

\item{max_points}{Defines the maximum points per batch.}

\item{use_integers}{Should integers (instead of doubles) be written if present?}
}
\value{
A list of server responses.
}
\description{
Write an xts object to an influxdb server
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/guides/writing_data/}
}
\seealso{
\code{\link[xts]{xts}}, \code{\link[influxdbr]{influx_connection}}
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
