% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb_explore_schema.R
\name{show_databases}
\alias{show_databases}
\title{show_databases}
\usage{
show_databases(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\value{
A character vector containing the database names.
}
\description{
Show databases
}
\details{
This function is a convenient wrapper for showing all databases
by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/query_language/schema_exploration.html}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

