% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_diagnostics.R
\name{show_diagnostics}
\alias{show_diagnostics}
\title{Show diagnostics}
\usage{
show_diagnostics(con)
}
\arguments{
\item{con}{An \code{influx_connection} object (s. \code{\link{influx_connection}}).}
}
\value{
A tibble with diagnostics.
}
\description{
This function is calls \code{influx_query} to receive some diagnostics.
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
