% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_diagnostics.R
\name{show_diagnostics}
\alias{show_diagnostics}
\title{Show diagnostics}
\usage{
show_diagnostics(con)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}
}
\value{
A list with diagnostics.
}
\description{
This function is a convenient wrapper for showing diagnostics
by calling \code{influx_query} with the corresponding query.
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
