% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_field_keys}
\alias{show_field_keys}
\title{show_field_keys}
\usage{
show_field_keys(con, db, from = NULL)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{from}{Query a specific measurement.}
}
\value{
A list with data.frames containing field keys.
}
\description{
Show field keys
}
\details{
This function is a convenient wrapper for showing the
field keys across each measurement in the database by calling
\code{influx_query} with the corresponding query.
A measurement can be specified (\code{from}).
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/query_language/schema_exploration/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

