% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_authentication_and_authorization.R
\name{show_grants}
\alias{show_grants}
\title{show_grants}
\usage{
show_grants(con, user)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{user}{The username.}
}
\description{
Show grants
}
\details{
This function is a convenient wrapper for listing a user's database privileges
by calling \code{influx_query} with the corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/administration/authentication_and_authorization/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

