% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_retention_policies}
\alias{show_retention_policies}
\title{show_retention_policies}
\usage{
show_retention_policies(con, db)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}
}
\value{
A data.frame containing the retention plocies.
}
\description{
Show retention policies
}
\details{
This function is a convenient wrapper for listing the existent retention
policies on a given database by calling \code{influx_query} with the
corresponding query.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/query_language/schema_exploration/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

