% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_explore_schema.R
\name{show_series}
\alias{show_series}
\title{show_series}
\usage{
show_series(con, db, from = NULL, where = NULL)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{from}{Query a specific measurement.}

\item{where}{Apply filter on tag key values.}
}
\value{
A list of data.frame objects.
}
\description{
Show series
}
\details{
This function is a convenient wrapper for showing all series with distinct
key-value pairs by calling \code{influx_query} with
the corresponding query. Series can be shown for a specific meausurement with
the (\code{from}) parameter and can also be filtered by tag key values with
the \code{where} clause.
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/query_language/schema_exploration.html}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

