% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_series}
\alias{show_series}
\title{show_series}
\usage{
show_series(con, db, measurement = NULL, where = NULL)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Query a specific measurement.}

\item{where}{Apply filter on tag key values.}
}
\value{
A list of data.frame objects.
}
\description{
Show series
}
\details{
This function is a convenient wrapper for showing all series with distinct
key-value pairs by calling \code{influx_query} with
the corresponding query. Series can be shown for a specific meausurement with
the (\code{measurement}) parameter and can also be filtered by tag key values with
the \code{where} clause.
}
\references{
\url{https://docs.influxdata.com/influxdb/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
