% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_diagnostics.R
\name{show_stats}
\alias{show_stats}
\title{Show stats}
\usage{
show_stats(con)
}
\arguments{
\item{con}{An \code{influx_connection} object (s. \code{\link{influx_connection}}).}
}
\value{
A tibble.
}
\description{
This function is calls \code{influx_query} to receive some stats.
}
\section{Warning}{

InfluxDB response might take some time.
}

\seealso{
\code{\link[influxdbr]{influx_connection}}
}
