% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_schema_exploration.R
\name{show_tag_keys}
\alias{show_tag_keys}
\title{show_tag_keys}
\usage{
show_tag_keys(con, db, measurement = NULL)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{measurement}{Query a specific measurement.}
}
\value{
A list of character vectors containing tag keys.
}
\description{
Show tag keys
}
\details{
This function is a convenient wrapper for showing all unique tag keys
associated with each measurement by calling
\code{influx_query} with the corresponding query.
The query can include a measurement (\code{measurement}) and tag key value (\code{where})
conditions, so only certain tag keys are shown.
}
\references{
\url{https://docs.influxdata.com/influxdb/v0.10/query_language/schema_exploration/}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
