% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/influxdb_explore_schema.R
\name{show_tag_values}
\alias{show_tag_values}
\title{show_series}
\usage{
show_tag_values(con, db, from = NULL, key)
}
\arguments{
\item{con}{An influx_connection object (s. \code{influx_connection}).}

\item{db}{Sets the target database for the query.}

\item{from}{Query a specific measurement.}

\item{key}{The key to be queried.}
}
\value{
A character vector containing tag values.
}
\description{
Show tag values
}
\details{
This function is a convenient wrapper for showing the unique set of
tag values for each measurement, for a given tag key by calling
\code{influx_query} with the corresponding query.
Tag values can be filtered by a specific measurement (\code{from}).
}
\author{
Dominik Leutnant (\email{leutnant@fh-muenster.de})
}
\references{
\url{https://influxdb.com/docs/v0.9/query_language/schema_exploration.html}
}
\seealso{
\code{\link[influxdbr]{influx_connection}}
}

